ported volume plugin
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2059 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/volume/VolumePlugin.h
@@ -62,7 +62,7 @@ private:
     QStringList m_params;
 
     /** amplification factor */
-    double m_factor;
+    float m_factor;
 
     /** mode for amplification selection */
     int m_mode;
